all data is coming from database now

+++ src/Entity/OffersMenus.php
@@ -33,7 +33,7 @@ class OffersMenus
     private $subtitle;
 
     /**
-     * @ORM\Column(type="integer")
+     * @ORM\Column(type="float")
      */
     private $price;
 
@@ -83,12 +83,12 @@ class OffersMenus
         return $this;
     }
 
-    public function getPrice(): ?int
+    public function getPrice(): ?float
     {
         return $this->price;
     }
 
-    public function setPrice(int $price): self
+    public function setPrice(float $price): self
     {
         $this->price = $price;
